delted unwanted file

+++ utils/surface_functions.py
@@ -103,15 +103,15 @@ def n_or_more_neg(D, teacher, rad, student, T, n, lr_1, lr_2, steps,experiment_p
     Q += dt * dQ
 
     if step % D == 0:
-      data['r'][step/D -1] = R
-      data['q'][step/D-1] = Q
+      data['r'][int(step/D) -1] = R
+      data['q'][int(step/D)-1] = Q
       
       p_correct = p_T_correct(Q,R,1)
       P = 0
       for i in range(n,T+1):
         P += scipy.special.binom(T,i) * p_correct**i * (1-p_correct)**(T-i)
       
-      data['p'][step/D -1] = P
+      data['p'][int(step/D) -1] = P
 
     step += 1
 
@@ -163,11 +163,11 @@ def all_neg(D, teacher, student, T, lr_1, lr_2, steps, experiment_path):
     Q += dt * dQ
 
     if step % D == 0:
-      data['r'][step/D -1] = R
-      data['q'][step/D -1] = Q
+      data['r'][int(step/D) -1] = R
+      data['q'][int(step/D) -1] = Q
       
       P = p_T_correct(Q,R,T)
-      data['p'][step/D -1] = P
+      data['p'][int(step/D) -1] = P
 
     step += 1
